NETOBSERV-2307: NETOBSERV-2315: fix several IPFIX issues (#1019)
* NETOBSERV-2307: NETOBSERV-2315: fix several IPFIX issues

- Fixed flows without ports that generated errors in logs, and were not exported. It
  shouldn't matter that ports are missing (e.g. ICMP)
- More generally, any missing field won't trigger an error anymore
- Some fields were missing: icmp type/code, tcp flags
- Fix resending templates in case of collector being restarted

* fix sending decoded tcpflags on ipfix

* Use patched go-ipfix, add write+read test...

The go-ipfix patch regularly recheck the udp connection, e.g. to account
for restarted host with a different resolved IP

... also:
- allow to provide data mapping for ingesting with goflow2
- rename API collector => ipfix
- simplify ingest API with more default values
- write ipfix: do not resend templates every second! Instead, use
  configurable periodicity

* Add simple example of ipfix ingester

* Warn about deprecated API usage

* fix race

+++ pkg/test/ipfix.go
@@ -26,7 +26,7 @@ type IPFIXClient struct {
 }
 
 // NewIPFIXClient returns an IPFIXClient that sends data to the given port
-func NewIPFIXClient(port int) (*IPFIXClient, error) {
+func NewIPFIXClient(port uint) (*IPFIXClient, error) {
 	conn, err := net.Dial("udp", fmt.Sprintf(":%d", port))
 	if err != nil {
 		return nil, fmt.Errorf("can't open UDP connection on port %d :%w",
@@ -100,7 +100,7 @@ func (ke *IPFIXClient) sendMessage(set entities.Set) error {
 }
 
 // UDPPort asks the kernel for a free open port that is ready to use.
-func UDPPort() (int, error) {
+func UDPPort() (uint, error) {
 	addr, err := net.ResolveUDPAddr("udp", "127.0.0.1:0")
 	if err != nil {
 		return 0, err
@@ -111,5 +111,5 @@ func UDPPort() (int, error) {
 		return 0, err
 	}
 	defer l.Close()
-	return l.LocalAddr().(*net.UDPAddr).Port, nil
+	return uint(l.LocalAddr().(*net.UDPAddr).Port), nil
 }
